move instantiation to interface
